vitos-v1/packages/vitos-monitor/ai/pyproject.toml
[project]
name = "vitos-ai"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
  "scikit-learn>=1.4",
  "numpy>=1.26",
  "httpx>=0.27",
  "pyyaml>=6",
  "click>=8.1",
]

[project.scripts]
vitos-ai = "vitos_ai.service:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
